refactored resize method for ie/edge
